examples_adk/requirements.txt
google-generativeai>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
